refactor: enhance GUI rendering and data processing logic
the x plot ranges seem incorrect still

- Added accumulation count display in the GUI.
- Improved axis limits for eye plot visualization.
- Refactored data processing to handle pre-trigger and post-trigger windows.
- Updated accumulation buffer management and statistics calculation.
- Introduced new header file for GUI rendering functions.

+++ reader/shared_data_types.hpp
@@ -40,9 +40,9 @@ struct DisplayData {
 
   // Metadata
   int original_sensor_index = -1;
-  size_t accumulation_count = 0;
-  int window_before_ms = 50; // Use a non-zero default
-  int window_after_ms = 150; // Use a non-zero default
+  int accumulation_count = 0; // How many traces were used for this data
+  int window_before_ms = 50;  // Use a non-zero default
+  int window_after_ms = 150;  // Use a non-zero default
 
   void clear() {
     x_data.clear();
